39-autogen_agent_chat_roundrobin/pyproject.toml
[project]
name = "autogen-agent-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "autogen-ext>=0.7.5",
    "autogen-agentchat>=0.0.1",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.5.0",
    "ipython>=9.10.0",
    "langchain-google-community>=3.0.5",
    "langchain>=1.2.10",
]
